Remove the redundancy in comparing floating-point values (#797)
The canonicalization was first introduced as part of https://github.com/sqlancer/sqlancer/pull/497 and then copied to subsequent implementations.
