+++ src/sqlancer/databend/test/tlp/DatabendQueryPartitioningGroupByTester.java
@@ -42,7 +42,7 @@ public class DatabendQueryPartitioningGroupByTester extends DatabendQueryPartiti
         List<String> secondResultSet = ComparatorHelper.getCombinedResultSetNoDuplicates(firstQueryString,
                 secondQueryString, thirdQueryString, combinedString, true, state, errors);
         ComparatorHelper.assumeResultSetsAreEqual(resultSet, secondResultSet, originalQueryString, combinedString,
-                state, DatabendQueryPartitioningBase::canonicalizeResultValue);
+                state, ComparatorHelper::canonicalizeResultValue);
     }
 
     @Override
