+++ src/sqlancer/doris/oracle/tlp/DorisQueryPartitioningBase.java
@@ -41,22 +41,6 @@ public class DorisQueryPartitioningBase extends TernaryLogicPartitioningOracleBa
         DorisErrors.addInsertErrors(errors);
     }
 
-    public static String canonicalizeResultValue(String value) {
-        if (value == null) {
-            return value;
-        }
-
-        switch (value) {
-        case "-0.0":
-            return "0.0";
-        case "-0":
-            return "0";
-        default:
-        }
-
-        return value;
-    }
-
     @Override
     public void check() throws SQLException {
         s = state.getSchema();
