+++ src/sqlancer/doris/oracle/tlp/DorisQueryPartitioningHavingTester.java
@@ -53,7 +53,7 @@ public class DorisQueryPartitioningHavingTester extends DorisQueryPartitioningBa
         List<String> secondResultSet = ComparatorHelper.getCombinedResultSet(firstQueryString, secondQueryString,
                 thirdQueryString, combinedString, !orderBy, state, errors);
         ComparatorHelper.assumeResultSetsAreEqual(resultSet, secondResultSet, originalQueryString, combinedString,
-                state, DorisQueryPartitioningBase::canonicalizeResultValue);
+                state, ComparatorHelper::canonicalizeResultValue);
     }
 
     @Override
