+++ src/sqlancer/doris/oracle/tlp/DorisQueryPartitioningWhereTester.java
@@ -47,7 +47,7 @@ public class DorisQueryPartitioningWhereTester extends DorisQueryPartitioningBas
         List<String> secondResultSet = ComparatorHelper.getCombinedResultSet(firstQueryString, secondQueryString,
                 thirdQueryString, combinedString, !orderBy, state, errors);
         ComparatorHelper.assumeResultSetsAreEqual(resultSet, secondResultSet, originalQueryString, combinedString,
-                state, DorisQueryPartitioningBase::canonicalizeResultValue);
+                state, ComparatorHelper::canonicalizeResultValue);
     }
 
 }
