+++ src/sqlancer/duckdb/test/DuckDBQueryPartitioningDistinctTester.java
@@ -38,7 +38,7 @@ public class DuckDBQueryPartitioningDistinctTester extends DuckDBQueryPartitioni
         List<String> secondResultSet = ComparatorHelper.getCombinedResultSetNoDuplicates(firstQueryString,
                 secondQueryString, thirdQueryString, combinedString, true, state, errors);
         ComparatorHelper.assumeResultSetsAreEqual(resultSet, secondResultSet, originalQueryString, combinedString,
-                state, DuckDBQueryPartitioningBase::canonicalizeResultValue);
+                state, ComparatorHelper::canonicalizeResultValue);
     }
 
 }
