+++ src/sqlancer/duckdb/test/DuckDBQueryPartitioningGroupByTester.java
@@ -41,7 +41,7 @@ public class DuckDBQueryPartitioningGroupByTester extends DuckDBQueryPartitionin
         List<String> secondResultSet = ComparatorHelper.getCombinedResultSetNoDuplicates(firstQueryString,
                 secondQueryString, thirdQueryString, combinedString, true, state, errors);
         ComparatorHelper.assumeResultSetsAreEqual(resultSet, secondResultSet, originalQueryString, combinedString,
-                state, DuckDBQueryPartitioningBase::canonicalizeResultValue);
+                state, ComparatorHelper::canonicalizeResultValue);
     }
 
     @Override
