+++ src/sqlancer/duckdb/test/DuckDBQueryPartitioningHavingTester.java
@@ -47,7 +47,7 @@ public class DuckDBQueryPartitioningHavingTester extends DuckDBQueryPartitioning
         List<String> secondResultSet = ComparatorHelper.getCombinedResultSet(firstQueryString, secondQueryString,
                 thirdQueryString, combinedString, !orderBy, state, errors);
         ComparatorHelper.assumeResultSetsAreEqual(resultSet, secondResultSet, originalQueryString, combinedString,
-                state, DuckDBQueryPartitioningBase::canonicalizeResultValue);
+                state, ComparatorHelper::canonicalizeResultValue);
     }
 
     @Override
