+++ src/sqlancer/questdb/test/QuestDBQueryPartitioningBase.java
@@ -3,7 +3,6 @@ package sqlancer.questdb.test;
 import java.sql.SQLException;
 import java.util.ArrayList;
 import java.util.List;
-import java.util.Objects;
 import java.util.stream.Collectors;
 
 import sqlancer.Randomly;
@@ -46,15 +45,6 @@ public class QuestDBQueryPartitioningBase
                     .collect(Collectors.toList());
         }
         return columns;
-    }
-
-    public static String canonicalizeResultValue(String value) {
-        // Rule: -0.0 should be canonicalized to 0.0
-        if (Objects.equals(value, "-0.0")) {
-            return "0.0";
-        }
-
-        return value;
     }
 
     @Override
