+++ src/sqlancer/questdb/test/QuestDBQueryPartitioningWhereTester.java
@@ -36,6 +36,6 @@ public class QuestDBQueryPartitioningWhereTester extends QuestDBQueryPartitionin
         List<String> secondResultSet = ComparatorHelper.getCombinedResultSet(firstQueryString, secondQueryString,
                 thirdQueryString, combinedString, false, state, errors);
         ComparatorHelper.assumeResultSetsAreEqual(resultSet, secondResultSet, originalQueryString, combinedString,
-                state, QuestDBQueryPartitioningBase::canonicalizeResultValue);
+                state, ComparatorHelper::canonicalizeResultValue);
     }
 }
